Move JPetTaskFactory to a separate folder
Also, change include path in JPetManager
